+++ src/biz/bokhorst/xprivacy/Meta.java
@@ -98,9 +98,10 @@ public class Meta {
 		mListHook.add(new Hook("identification", "GservicesProvider", "com.google.android.providers.gsf.permission.READ_GSERVICES,com.google.android.providers.gsf.permission.WRITE_GSERVICES", 1, null, null).dangerous());
 		mListHook.add(new Hook("identification", "SERIAL", "", 1, null, null).restart().noUsageData());
 
-		mListHook.add(new Hook("internet", "getAllByName", "INTERNET", 1, null, null));
-		mListHook.add(new Hook("internet", "getByAddress", "INTERNET", 1, null, null));
-		mListHook.add(new Hook("internet", "getByName", "INTERNET", 1, null, null));
+		mListHook.add(new Hook("internet", "getAllByName", "INTERNET", 1, "0.0", null).dangerous());
+		mListHook.add(new Hook("internet", "getByAddress", "INTERNET", 1, "0.0", null).dangerous());
+		mListHook.add(new Hook("internet", "getByName", "INTERNET", 1, "0.0", null).dangerous());
+
 		mListHook.add(new Hook("internet", "getByInetAddress", "INTERNET", 1, null, null));
 		mListHook.add(new Hook("internet", "getNetworkInterfaces", "INTERNET", 1, null, null));
 		mListHook.add(new Hook("internet", "inet", "INTERNET", 1, null, null).dangerous().restart().noUsageData());